refactor(models): extract routing + spend tabs into components (phase 2a)
Second phase of the models-page.tsx split — the first two tab extractions,
fed by the phase-1 data hook.

- use-models-data.ts exports `ModelsData = ReturnType<typeof useModelsData>`
  for the tab props.
- routing-tab.tsx (RoutingTab) and spend-tab.tsx (SpendTab) each take
  `{ m }: { m: ModelsData }` and destructure what they need. The
  routing-only constants (ROUTING_ORIGINS, THINKING_LEVELS) move into
  routing-tab; the spend-only SPEND_WINDOWS + formatUsdMicros move into
  spend-tab.

models-page.tsx switches to `const m = useModelsData()` + a slimmer
destructure for the still-inline agents/available/aliases tabs, and renders
`<RoutingTab m={m} />` / `<SpendTab m={m} />`. 804 -> 561.

Pure behavior-preserving relocation — the tab JSX is byte-identical, now in
presentational components reading one props object.

Verified: typecheck + eslint + models-page.test (10/0, renders all tabs incl.
routing + spend as components) + full suite (5124/0) + binary build +
isolated-home boot smoke (Models plugin loads; rebuilt client.js → 200).
